requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
click==8.2.1
distro==1.9.0
dnspython==2.7.0
fastapi==0.116.1
joblib==1.5.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.102.0
pydantic==2.11.7
pydantic-core==2.33.2
pymongo==4.14.1
python-dotenv==1.1.1
pyyaml==6.0.2
requests==2.32.5
scikit-learn==1.7.2
sniffio==1.3.1
starlette==0.47.3
transformers==4.56.1
torch==2.8.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.1
uvicorn==0.35.0
uvloop==0.21.0
velebit-useful-logs==1.0.3
watchfiles==1.1.0
websockets==15.0.1
